Converted join house section to independent file
