+++ client/src/components/ViewHouse.tsx
@@ -1,8 +1,5 @@
-import React, { useEffect, useState, Fragment } from "react";
-import { useFormik } from "formik";
-import * as Yup from "yup";
+import React, { useState, Fragment } from "react";
 import { Link } from "react-router-dom";
-import { Select, SelectChangeEvent, MenuItem, InputLabel } from "@mui/material";
 
 import Add from "@mui/icons-material/AddCircleOutline";
 import Delete from "@mui/icons-material/HighlightOff";
@@ -10,7 +7,6 @@ import Leave from "@mui/icons-material/ExitToApp";
 import VerifiedIcon from "@mui/icons-material/Verified";
 import DoNotDisturbIcon from "@mui/icons-material/DoNotDisturb";
 import "./viewHouse.scss";
-import CreateHouse from "./CreateHouse";
 
 interface Values {
   house_name: String;
@@ -18,71 +14,19 @@ interface Values {
   members: [];
 }
 
-function ViewHouse() {
-  const [house, setHouse] = useState("");
-  const [hasHouse, setHasHouse] = useState(true);
-  const [joinHouse, setJoinHouse] = useState(true);
-  const [options, setOptions] = useState([]);
+interface Props {
+  setHasHouse: React.Dispatch<React.SetStateAction<boolean>>;
+  setJoinHouse: React.Dispatch<React.SetStateAction<boolean>>;
+  values: Values;
+}
+
+const ViewHouse = (props: Props) => {
+  const { setHasHouse, setJoinHouse, values } = props;
   const [issue, setIssue] = useState("");
-  const [values, setValues] = useState<Values>({
-    house_name: "",
-    admin: false,
-    members: [],
-  });
 
-  useEffect(() => {
-    fetch("/house", {
-      method: "get",
-    }).then((response) => {
-      console.log(values["house_name"]);
-      if (response.redirected) {
-        window.location.href = response.url;
-      } else if (response.status === 209) {
-        setHasHouse(false);
-        response.json().then((responseJson) => {
-          setOptions(responseJson);
-        });
-      } else {
-        response.json().then((responseJson) => {
-          setValues({
-            house_name: responseJson["house_name"],
-            admin: responseJson["admin"],
-            members: responseJson["members"],
-          });
-        });
-      }
-    });
-  }, []);
-
-  const formik_join = useFormik({
-    initialValues: {},
-
-    onSubmit: () => {
-      fetch("/house/join", {
-        method: "post",
-        headers: {
-          Accept: "application/json",
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          house: house,
-        }),
-      }).then((response) => {
-        if (response.redirected) {
-          console.log(response);
-          window.location.href = response.url;
-        } else {
-          response.json().then((responseJson) => {
-            console.log(responseJson["issue"]);
-            setIssue(responseJson["issue"]);
-          });
-        }
-      });
-    },
-  });
-
-  const handleHouseChange = (event: SelectChangeEvent) => {
-    setHouse(event.target.value as string);
+  const handleCreateClick = () => {
+    setHasHouse(false);
+    setJoinHouse(false);
   };
 
   const handleDeleteClick = () => {
@@ -113,151 +57,73 @@ function ViewHouse() {
     });
   };
 
-  const handleCreateClick = () => {
-    setHasHouse(false);
-    setJoinHouse(false);
-  };
-
   return (
-    <div className="household">
-      {hasHouse && (
-        <div className="has-house">
-          <h1>House Details</h1>
-          <p className={"issue " + (issue && "active")}>{issue}</p>
-          <div className="house-actions">
-            {values["admin"] && (
-              <Link
-                onClick={() => handleDeleteClick()}
-                to=""
-                className="action"
-              >
-                <Delete className="icon" />
-                <label>Delete</label>
-              </Link>
-            )}
-            <Link onClick={() => handleLeaveClick()} to="" className="action">
-              <Leave className="icon" />
-              <label>Leave</label>
-            </Link>
-            <Link onClick={() => handleCreateClick()} to="" className="action">
-              <Add className="icon" />
-              <label>Create</label>
-            </Link>
-          </div>
-          <div className="details">
-            <div className="label">
-              <label>Name : </label>
-              <label>Members : </label>
-            </div>
-            <div className="value">
-              <span>{values["house_name"]}</span>
-              <div className={"table_container"}>
-                <table className="table">
-                  <thead className="table_head">
-                    <tr className="top_row">
-                      <th>Name</th>
-                      <th className="bool">Admin</th>
-                      <th className="bool">Verified</th>
-                    </tr>
-                  </thead>
-                  <tbody className="table_body">
-                    {values["members"].map((member) => {
-                      return (
-                        <Fragment key={member["id"]}>
-                          <tr className="rows">
-                            <td className="primary">{member["username"]}</td>
-                            <td className="primary bool">
-                              {member["admin"] ? (
-                                <VerifiedIcon />
-                              ) : (
-                                <DoNotDisturbIcon />
-                              )}
-                            </td>
-                            <td className="primary bool">
-                              {member["verified"] ? (
-                                <VerifiedIcon />
-                              ) : (
-                                <DoNotDisturbIcon />
-                              )}
-                            </td>
-                          </tr>
-                        </Fragment>
-                      );
-                    })}
-                  </tbody>
-                </table>
-              </div>
-            </div>
+    <div className="has-house">
+      <h1>House Details</h1>
+      <p className={"issue " + (issue && "active")}>{issue}</p>
+      <div className="house-actions">
+        {values["admin"] && (
+          <Link onClick={() => handleDeleteClick()} to="" className="action">
+            <Delete className="icon" />
+            <label>Delete</label>
+          </Link>
+        )}
+        <Link onClick={() => handleLeaveClick()} to="" className="action">
+          <Leave className="icon" />
+          <label>Leave</label>
+        </Link>
+        <Link onClick={() => handleCreateClick()} to="" className="action">
+          <Add className="icon" />
+          <label>Create</label>
+        </Link>
+      </div>
+      <div className="details">
+        <div className="label">
+          <label>Name : </label>
+          <label>Members : </label>
+        </div>
+        <div className="value">
+          <span>{values["house_name"]}</span>
+          <div className={"table_container"}>
+            <table className="table">
+              <thead className="table_head">
+                <tr className="top_row">
+                  <th>Name</th>
+                  <th className="bool">Admin</th>
+                  <th className="bool">Verified</th>
+                </tr>
+              </thead>
+              <tbody className="table_body">
+                {values["members"].map((member) => {
+                  return (
+                    <Fragment key={member["id"]}>
+                      <tr className="rows">
+                        <td className="primary">{member["username"]}</td>
+                        <td className="primary bool">
+                          {member["admin"] ? (
+                            <VerifiedIcon />
+                          ) : (
+                            <DoNotDisturbIcon />
+                          )}
+                        </td>
+                        <td className="primary bool">
+                          {member["verified"] ? (
+                            <VerifiedIcon />
+                          ) : (
+                            <DoNotDisturbIcon />
+                          )}
+                        </td>
+                      </tr>
+                    </Fragment>
+                  );
+                })}
+              </tbody>
+            </table>
           </div>
         </div>
-      )}
-      {!hasHouse && (
-        <div className="popup">
-          <div className="content">
-            <button className="close" onClick={() => console.log("hi")}>
-              x
-            </button>
-            <div className="add-house-actions">
-              <Link onClick={() => setJoinHouse(true)} to="">
-                <span className={"action " + (joinHouse && "active")}>
-                  Join
-                </span>
-              </Link>
-              <Link onClick={() => setJoinHouse(false)} to="">
-                <span className={"action " + (!joinHouse && "active")}>
-                  Create
-                </span>
-              </Link>
-            </div>
-            <div className="forms">
-              {joinHouse && (
-                <div className="house join">
-                  <form onSubmit={formik_join.handleSubmit}>
-                    <div className="form">
-                      <p className={"formIssue " + (issue && "active")}>
-                        {issue}
-                      </p>
-                      <div className="fields">
-                        <div className="field">
-                          <InputLabel className={"label dropdown"}>
-                            {"Select House"}
-                          </InputLabel>
-                          <Select
-                            labelId={"label dropdown"}
-                            id={"id dropdown"}
-                            label={"House"}
-                            value={house}
-                            sx={{ maxHeight: 40, minWidth: 120 }}
-                            onChange={handleHouseChange}
-                            style={{ width: "60%" }}
-                          >
-                            {Object.keys(options).map((option) => {
-                              return (
-                                <MenuItem key={option} value={option}>
-                                  {options[parseInt(option)]}
-                                </MenuItem>
-                              );
-                            })}
-                          </Select>
-                        </div>
-
-                        <div className="button">
-                          <button disabled={!formik_join.isValid}>
-                            <span>Join</span>
-                          </button>
-                        </div>
-                      </div>
-                    </div>
-                  </form>
-                </div>
-              )}
-              {!joinHouse && <CreateHouse />}
-            </div>
-          </div>
-        </div>
-      )}
+      </div>
     </div>
   );
-}
+};
 
 export default ViewHouse;
